+++ src/importBarrelsRule.ts
@@ -1,10 +1,11 @@
 import * as fs from 'fs';
 import * as path from 'path';
-import { IRuleMetadata, RuleFailure, Rules, RuleWalker, Utils } from 'tslint/lib';
+import { Fix, IOptions, IRuleMetadata, Replacement, RuleFailure, Rules, RuleWalker, Utils } from 'tslint/lib';
 import { Expression, ImportDeclaration, SourceFile, StringLiteral, SyntaxKind } from 'typescript';
+import { CachedStat } from './utils/cachedStat';
 
 export class Rule extends Rules.AbstractRule {
-  public static metadata:IRuleMetadata = {
+  public static readonly metadata:IRuleMetadata = {
     ruleName: 'import-barrels',
     description: Utils.dedent`
       Enforces usage of barrels (\`index.ts\`) when importing from a directory that has a barrel file.`,
@@ -17,7 +18,9 @@ export class Rule extends Rules.AbstractRule {
       An argument object may be optionally provided, with the following properties:
       
       * \`noExplicitBarrels = false\`: disallows usage of explicitly named barrels in import statements (\`import foo from './foo/index'\`)
-      * \`fileExtensions = ['ts', 'js']\`: uses the provided file extensions for module and barrel file lookup`,
+      * \`fileExtensions = ['ts', 'js']\`: uses the provided file extensions for module and barrel file lookup
+      * \`fixWithExplicitBarrelImport\`: uses the provided string to replace non-barrel imports in \`--fix\` mode
+        (i.e. when set to \`'index'\`, \`import foo from './foo/some-module'\` becomes \`import foo from './foo/index'\`)`,
     optionExamples: ['[true, {"noExplicitBarrels": false, "fileExtensions": ["ts", "js"]}]'],
     options: {
       type: 'array',
@@ -34,6 +37,9 @@ export class Rule extends Rules.AbstractRule {
             },
             minLength: 1,
           },
+          fixWithExplicitBarrelImport: {
+            type: 'string',
+          },
         },
       },
       minLength: 1,
@@ -43,38 +49,66 @@ export class Rule extends Rules.AbstractRule {
     typescriptOnly: false,
   };
 
-  public static USE_BARREL_FAILURE_STRING = 'Use barrel (index) files for imports if they are available for path: ';
-  public static NO_EXPLICIT_BARRELS_FAILURE_STRING = "Don't import barrel files by name, import containing directory instead for path: ";
+  public static readonly USE_BARREL_FAILURE_STRING = 'Use barrel (index) files for imports if they are available for path: ';
+  public static readonly NO_EXPLICIT_BARRELS_FAILURE_STRING =
+    "Don't import barrel files by name, import containing directory instead for path: ";
+
+  public static readonly DEFAULT_OPTIONS = {
+    fileExtensions: ['ts', 'js'],
+    noExplicitBarrels: false,
+    fixWithExplicitBarrelImport: ''
+  };
 
   public apply(sourceFile:SourceFile):RuleFailure[] {
     return this.applyWithWalker(new ImportBarrelsWalker(sourceFile, this.getOptions()));
   }
 }
 
+enum CheckResult {
+  OK,
+  ExplicitBarrelsForbidden,
+  NonBarrelImport
+}
+
 class ImportBarrelsWalker extends RuleWalker {
-  private statCache = new Map<string, fs.Stats>();
+  private static readonly resultErrorMap = {
+    [CheckResult.NonBarrelImport]: Rule.USE_BARREL_FAILURE_STRING,
+    [CheckResult.ExplicitBarrelsForbidden]: Rule.NO_EXPLICIT_BARRELS_FAILURE_STRING,
+  };
+  private readonly cachedStat = new CachedStat();
+  private readonly ruleOptions:typeof Rule.DEFAULT_OPTIONS;
+
+  constructor(sourceFile:SourceFile, options:IOptions) {
+    super(sourceFile, options);
+
+    this.ruleOptions = Object.assign({}, Rule.DEFAULT_OPTIONS, this.getOptions()[0] || {});
+  }
 
   protected visitImportDeclaration(node:ImportDeclaration) {
     const moduleExpression = node.moduleSpecifier;
-    const moduleExpressionError = this.getModuleExpressionErrorMessage(moduleExpression);
+    const expressionCheckResult = this.checkModuleExpression(moduleExpression);
 
-    if (moduleExpressionError) {
-      this.addFailureAtNode(moduleExpression, moduleExpressionError);
+    if (expressionCheckResult !== CheckResult.OK) {
+      this.addFailureAtNode(
+        moduleExpression,
+        ImportBarrelsWalker.resultErrorMap[expressionCheckResult] + moduleExpression.getText(),
+        this.getFix(<StringLiteral>moduleExpression, expressionCheckResult)
+      );
     }
 
     super.visitImportDeclaration(node);
   }
 
-  private getModuleExpressionErrorMessage(expression:Expression):string|null {
+  private checkModuleExpression(expression:Expression):CheckResult {
     if (expression.kind !== SyntaxKind.StringLiteral) {
-      return null;
+      return CheckResult.OK;
     }
 
     const modulePathText = (<StringLiteral>expression).text;
 
     // check only relative paths
     if (!modulePathText.startsWith('.')) {
-      return null;
+      return CheckResult.OK;
     }
 
     const sourceFileRelative = this.getSourceFile().path;
@@ -85,80 +119,49 @@ class ImportBarrelsWalker extends RuleWalker {
     // enforce barrel usage only on files that are not in the same directory or in one of the sub-directories
     // of the module
     if (sourceFileDirAbsolute.startsWith(moduleDirAbsolute)) {
-      return null;
+      return CheckResult.OK;
     }
 
     const moduleStats = this.getModuleStats(moduleAbsolute);
 
     // only file imports are of interest
     if (!moduleStats || moduleStats.isDirectory()) {
-      return null;
+      return CheckResult.OK;
     }
 
     // if module's name is 'index', it must be an explicit barrel import, dirs were excluded earlier
     if (path.parse(moduleAbsolute).name === 'index') {
-      return this.getExplicitBarrelsAllowed() ? null : Rule.NO_EXPLICIT_BARRELS_FAILURE_STRING + expression.getText();
+      return this.ruleOptions.noExplicitBarrels ? CheckResult.ExplicitBarrelsForbidden : CheckResult.OK;
     }
 
-    return this.getModuleFileExtensions()
+    const dirHasBarrelFile = this.ruleOptions.fileExtensions
       .map(ext => path.join(moduleDirAbsolute, `index.${ext}`))
-      .some(file => this.isFile(file)) ? Rule.USE_BARREL_FAILURE_STRING + expression.getText() : null;
+      .some(file => this.cachedStat.isFile(file));
+
+    return dirHasBarrelFile ? CheckResult.NonBarrelImport : CheckResult.OK;
   }
 
   private getModuleStats(modulePath:string):fs.Stats|null {
-    let stats = this.cachedStatSync(modulePath);
+    const modulePathCandidates = [modulePath, ...this.ruleOptions.fileExtensions.map(suffix => `${modulePath}.${suffix}`)];
 
-    if (!stats) {
-      this.getModuleFileExtensions().some(suffix => {
-        stats = this.cachedStatSync(`${modulePath}.${suffix}`);
-        return stats !== null;
-      });
-    }
+    let stats:fs.Stats|null = null;
+    modulePathCandidates.some(modulePathCandidate => {
+      stats = this.cachedStat.statSync(modulePathCandidate);
+      return stats !== null;
+    });
 
     return stats;
   }
 
-  private getModuleFileExtensions():string[] {
-    let extensions = this.getOptionsObject().fileExtensions || [];
+  private getFix(moduleSpecifier:StringLiteral, checkResult:CheckResult):Fix {
+    let replacement = path.dirname(moduleSpecifier.text);
 
-    if (!extensions.length) {
-      extensions = ['ts', 'js'];
-    }
-    return extensions;
-  }
-
-  private getExplicitBarrelsAllowed():boolean {
-    const {noExplicitBarrels = false} = this.getOptionsObject();
-
-    return !noExplicitBarrels;
-  }
-
-  private getOptionsObject():{fileExtensions?:string[], noExplicitBarrels?:boolean} {
-    return this.getOptions()[0] || {};
-  }
-
-  private isFile(path:string):boolean {
-    const stats = this.cachedStatSync(path);
-    return stats != null && stats.isFile();
-  }
-
-  private cachedStatSync(path:string):fs.Stats|null {
-    if (this.statCache.has(path)) {
-      return <fs.Stats>this.statCache.get(path);
+    if (checkResult === CheckResult.NonBarrelImport && this.ruleOptions.fixWithExplicitBarrelImport) {
+      replacement += `/${this.ruleOptions.fixWithExplicitBarrelImport}`;
     }
 
-    const stat = this.statSync(path);
-    if (stat !== null) {
-      this.statCache.set(path, stat);
-    }
-    return stat;
-  }
-
-  private statSync(path:string):fs.Stats|null {
-    try {
-      return fs.statSync(path);
-    } catch (e) {
-      return null;
-    }
+    return new Fix(Rule.metadata.ruleName, [
+      // account for quotes
+      new Replacement(moduleSpecifier.getStart() + 1, moduleSpecifier.getWidth() - `''`.length, replacement)]);
   }
 }