+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Change Log
 v2.1.0
 ---
 * [new-fixer] `jasmine-no-lambda-expression-callbacks`
+* [new-fixer] `import-barrels`
+* [new-rule-option] `fixWithExplicitBarrelImport` added to `import-barrels`
 * [enhancement] Added preliminary checks for `jasmine-no-lambda-expression-callbacks` rule to skip walking file if there are no jasmine
 top-level statements.